digital-forensics/lab-1/requirements.txt
numpy~=2.1.1
scikit-image~=0.24.0
pillow~=10.4.0